moved facet sorting from IxTheo to TueFind
