+++ module/TueFind/src/TueFind/Search/Results/Factory.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace TueFind\Search\Results;
+use Zend\ServiceManager\ServiceManager;
+
+class Factory extends \VuFind\Search\Results\Factory
+{
+    /**
+     * Factory for Solr results object.
+     *
+     * @param ServiceManager $sm Service manager.
+     *
+     * @return \VuFind\Search\Solr\Results
+     */
+    public static function getSolr(ServiceManager $sm)
+    {
+        $factory = new PluginFactory();
+        $solr = $factory->createServiceWithName($sm, 'solr', 'Solr');
+        $config = $sm->getServiceLocator()
+            ->get('VuFind\Config')->get('config');
+        $spellConfig = isset($config->Spelling)
+            ? $config->Spelling : null;
+        $solr->setSpellingProcessor(
+            new \VuFind\Search\Solr\SpellingProcessor($spellConfig)
+        );
+        return $solr;
+    }
+}
